+++ index.js
@@ -27,6 +27,9 @@ exportOnDemand({
 	SnapshotPlugin: () => require('./plugins/SnapshotPlugin'),
 	VerboseLogPlugin: () => require('./plugins/VerboseLogPlugin'),
 	WebOSMetaPlugin: () => require('./plugins/WebOSMetaPlugin'),
+	ViteHtmlPlugin: () => require('./plugins/ViteHtmlPlugin'),
+	ViteILibPlugin: () => require('./plugins/ViteILibPlugin'),
+	ViteWebOSMetaPlugin: () => require('./plugins/ViteWebOSMetaPlugin')
 	EsbuildHtmlPlugin: () => require('./plugins/EsbuildHtmlPlugin'),
 	EsbuildILibPlugin: () => require('./plugins/EsbuildILibPlugin'),
 	EsbuildWebOSMetaPlugin: () => require('./plugins/EsbuildWebOSMetaPlugin')